more lottery stuff idk

+++ cogs/dankmemer/lottery.py
@@ -1,5 +1,6 @@
 import contextlib
-from typing import Literal
+import random
+from typing import Literal, Optional
 
 import discord
 from discord import SlashCommandGroup
@@ -10,6 +11,78 @@ from utils.buttons import confirm
 from utils.context import DVVTcontext
 
 accepted_lottery_types = ['dank', 'karuta', 'owo']
+
+class GoldPotTemplateModal(discord.ui.Modal):
+    def __init__(self, client, lottery_type: Literal['dank', 'karuta', 'owo'], lotto_entries, winner:Optional[discord.User] = None, default_winner_option: Optional[str] = None, default_prize: Optional[str] = None):
+        self.client = client
+        self.winner = winner
+        self.default_winner_option = default_winner_option
+        self.default_prize = default_prize
+        super().__init__(title="Generate Gold Pot Template", custom_id=f"lotto-template:{lottery_type}:{lotto_entries}")
+        if self.winner is not None:
+            if self.default_winner_option is not None:
+                self.add_item(discord.ui.InputText(label="Lottery Winner",
+                                                   placeholder="MUST be a Discord User ID. Leave this blank if user can't be found.",
+                                                   required=False, value=self.default_winner_option,
+                                                   style=discord.InputTextStyle.short))
+            else:
+                self.add_item(discord.ui.InputText(label="Lottery Winner", placeholder="MUST be a Discord User ID. Leave this blank if user can't be found.", required=False, value=str(winner.id), style=discord.InputTextStyle.short))
+        if self.default_prize is not None:
+            self.add_item(discord.ui.InputText(label="Lottery Prize", required=False, value=default_prize, style=discord.InputTextStyle.short))
+        else:
+            self.add_item(discord.ui.InputText(label="Lottery Prize", style=discord.InputTextStyle.short))
+
+    async def callback(self, interaction: discord.Interaction):
+        remarks = []
+        lottery_type, lotto_entries = self.custom_id.split(':')[1], self.custom_id.split(':')[2]
+        try:
+            winner_id = int(self.children[0].value)
+        except ValueError:
+            if len(self.children[0].value) == 0:
+                user = None
+            else:
+                remarks.append(f"<:DVB_False:887589731515392000> Invalid User: `{self.children[0].value}`")
+                user = None
+        else:
+            if (user := self.client.get_user(winner_id)) is None:
+                remarks.append(f"<:DVB_False:887589731515392000> Invalid User: `{self.children[0].value}`")
+
+        prize = self.children[1].value
+        if lottery_type == 'owo':
+            emoji = ('<a:dv_fabledGemOwO:859773160357232651>', '<:dv_newOwO:859772980875886602>')
+        elif lottery_type == 'karuta':
+            emoji = ('<a:dv_bunbunDanceOwO:837749889496514570>', '<a:dv_aFishieOwO:837741744253829190>')
+        else:
+            emoji = ('<a:dv_pepeHypeOwO:837712517261688832>', '<:dv_panMoneyOwO:837713209796460564>')
+        embeds = []
+        if user is not None:
+            us_display = f"**{user.name}** ({user.mention})"
+        else:
+            us_display = "**USERNAME** (<@ID>)"
+        embeds.append(discord.Embed(title="Coin Prize Template", description=f"```\n{emoji[0]} **{lotto_entries}** users entered, {us_display} snagged an easy {emoji[1]} **{prize}** {emoji[1]}\n```", color=0xFFD700))
+        if lottery_type == 'dank':
+            embeds.append(discord.Embed(title="Item Prize Template", description=f"```\n{emoji[0]} **{lotto_entries}** users entered, {us_display} easily snagged **{prize}** worth {emoji[1]} **⏣ ** {emoji[1]}\n```", color=0xFFD700))
+        if len(remarks) > 0:
+            remarks = "**Warning**:\n\n" + "\n".join(remarks)
+            await interaction.response.send_message(remarks, embeds=embeds)
+        else:
+            await interaction.response.send_message(embeds=embeds)
+
+
+class LottoTemplateView(discord.ui.View):
+    def __init__(self, client, lottery_type, lotto_entries, winner):
+        self.client = client
+        self.lottery_type = lottery_type
+        self.lotto_entries = lotto_entries
+        self.winner = winner
+        super().__init__(timeout=None)
+
+    @discord.ui.button(label="Generate Lottery Win Template")
+    async def generate_lottery_template(self, button:discord.ui.Button, interaction: discord.Interaction):
+        await interaction.response.send_modal(GoldPotTemplateModal(self.client, self.lottery_type, self.lotto_entries, self.winner))
+
+
+
 
 def get_lotto_channel(guild_id: int, type_of_lottery: Literal['dank', 'karuta', 'owo']) -> int:
     if guild_id == 595457764935991326:
@@ -61,6 +134,28 @@ class Lottery(commands.Cog):
     def __init__(self, client):
         self.client: dvvt = client
 
+    async def check_non_hoster_consent(self, ctx: DVVTcontext, lottery_hoster: int):
+        if ctx.author.id != lottery_hoster:
+            us = self.client.get_user(lottery_hoster)
+            us = f"{us}" if us else str(lottery_hoster)
+            confirmview = confirm(ctx, self.client, 30.0)
+            confirmembed = discord.Embed(title="You are not the host of this lottery.", description=f"{us} is the host of this lottery, are you sure you want to continue?", color=discord.Color.orange())
+            confirmview.response = await ctx.send(embed=confirmembed, view=confirmview)
+            await confirmview.wait()
+            if confirmview.returning_value is not True:
+                confirmembed.color=discord.Color.red()
+                await confirmview.response.edit(embed=confirmembed)
+                return False
+            else:
+                confirmembed.color = discord.Color.green()
+                await confirmview.response.edit(embed=confirmembed)
+                return True
+        else:
+            return True
+
+
+
+
     @checks.has_permissions_or_role(manage_roles=True)
     @commands.group(name="lottery", invoke_without_command=True)
     async def lottery(self, ctx: DVVTcontext, lottery_id: int = None, users: commands.Greedy[discord.Member] = None):
@@ -72,12 +167,14 @@ class Lottery(commands.Cog):
         """
         if lottery_id is None:
             return await ctx.help()
-        if not (lottery_obj := await self.client.db.fetchrow("SELECT * FROM lotteries WHERE lottery_id = $1", lottery_id)):
-            await ctx.send(f"A lottery with the ID {lottery_id} doesn't exist.")
+        if not (lottery_obj := await self.client.db.fetchrow("SELECT * FROM lotteries WHERE lottery_id = $1 AND active = $2", lottery_id, True)):
+            await ctx.send(f"<:DVB_False:887589731515392000> An active lottery with the ID {lottery_id} doesn't exist.")
             return
         type_of_lottery = lottery_obj.get('lottery_type')
         if type_of_lottery not in ['dank', 'karuta', 'owo']:
             return await ctx.send("Invalid lottery type.")
+
+        await self.check_non_hoster_consent(ctx, lottery_obj.get('starter_id'))
 
         last_lottery_number = await self.client.db.fetchval("SELECT lottery_number FROM lottery_entries WHERE lottery_id = $1 ORDER BY lottery_number DESC LIMIT 1", lottery_id)
         if last_lottery_number is None:
@@ -210,8 +307,8 @@ class Lottery(commands.Cog):
         """
         if lottery_id is None:
             return await ctx.help()
-        if (lottery_db := await self.client.db.fetchrow("SELECT * FROM lotteries WHERE lottery_id=$1", lottery_id)) is None:
-            return await ctx.send(f"No lottery found with ID `{lottery_id}`.")
+        if (lottery_db := await self.client.db.fetchrow("SELECT * FROM lotteries WHERE lottery_id=$1 AND active = $2", lottery_id, True)) is None:
+            return await ctx.send(f"<:DVB_False:887589731515392000> An active lottery with the ID {lottery_id} does not exist.")
         if user is None:
             embed = discord.Embed(title=f"Reserved entries for {lottery_db.get('lottery_type')} lottery #{lottery_id}", color=self.client.embed_color)
             entries_str = []
@@ -224,18 +321,20 @@ class Lottery(commands.Cog):
             embed.description = "\n".join(entries_str)
             embed.set_footer(text="To remove a reserved number, run the command again like you would add the entry.")
             return await ctx.send(embed=embed)
-        if lottery_number is None:
-            return await ctx.send(f"You must specify a lottery entry number to reserve (or remove).")
-        if (num_reserved_entry := await self.client.db.fetchrow("SELECT * FROM reserved_entries WHERE lottery_id=$1 AND lottery_number = $2", lottery_id, lottery_number)) is not None:
-            if num_reserved_entry.get('lottery_user') != user.id:
-                us = self.client.get_user(num_reserved_entry.get('lottery_user'))
-                us = f"{us} ({us.id})" if us is not None else num_reserved_entry.get('lottery_user')
-                return await ctx.send(f"The lottery number `{lottery_number}` is already reserved for **{us}**.")
-            await self.client.db.execute("DELETE FROM reserved_entries WHERE lottery_id=$1 AND lottery_user=$2 AND lottery_number = $3", lottery_id, user.id, lottery_number)
-            return await ctx.send(f"Removed **{user}**'s `{lottery_number}` entry from the lottery's reserved entries.")
-        else:
-            await self.client.db.execute("INSERT INTO reserved_entries(lottery_id, lottery_user, lottery_number) VALUES($1, $2, $3)", lottery_id, user.id, lottery_number)
-            return await ctx.send(f"Added **{user}**'s `{lottery_number}` entry to the lottery's reserved entries.")
+
+        if await self.check_non_hoster_consent(ctx, lottery_db.get('starter_id')) is True:
+            if lottery_number is None:
+                return await ctx.send(f"You must specify a lottery entry number to reserve (or remove).")
+            if (num_reserved_entry := await self.client.db.fetchrow("SELECT * FROM reserved_entries WHERE lottery_id=$1 AND lottery_number = $2", lottery_id, lottery_number)) is not None:
+                if num_reserved_entry.get('lottery_user') != user.id:
+                    us = self.client.get_user(num_reserved_entry.get('lottery_user'))
+                    us = f"{us} ({us.id})" if us is not None else num_reserved_entry.get('lottery_user')
+                    return await ctx.send(f"The lottery number `{lottery_number}` is already reserved for **{us}**.")
+                await self.client.db.execute("DELETE FROM reserved_entries WHERE lottery_id=$1 AND lottery_user=$2 AND lottery_number = $3", lottery_id, user.id, lottery_number)
+                return await ctx.send(f"Removed **{user}**'s `{lottery_number}` entry from the lottery's reserved entries.")
+            else:
+                await self.client.db.execute("INSERT INTO reserved_entries(lottery_id, lottery_user, lottery_number) VALUES($1, $2, $3)", lottery_id, user.id, lottery_number)
+                return await ctx.send(f"Added **{user}**'s `{lottery_number}` entry to the lottery's reserved entries.")
 
 
     @checks.has_permissions_or_role(manage_roles=True)
@@ -245,23 +344,69 @@ class Lottery(commands.Cog):
         If you messed up, use this command to change the lottery's current entry number.
         This number should be the last number entered correctly.
         """
-        confirmview = confirm(ctx, self.client, 30.0)
-        confirmembed = discord.Embed(description=f"Are you sure you want to change the current entry number of **lottery #{lottery_id}** to `{count}`?\n\nThe last correctly entered entry will be `{count}`.\nThe next entry to be entered will be `{count+1}`, ", color=discord.Color.orange())
-        confirmview.response = await ctx.send(embed=confirmembed, view=confirmview)
-        await confirmview.wait()
-        if confirmview.returning_value is not True:
-            confirmembed.color = discord.Color.red()
+        if (lotto_object := await self.client.db.fetchrow("SELECT * FROM lotteries WHERE lottery_id = $1 AND active = $2", lottery_id, True)) is None:
+            return await ctx.send(f"<:DVB_False:887589731515392000> An active lottery with the ID {lottery_id} does not exist.")
+
+        if await self.check_non_hoster_consent(ctx, lotto_object.get('starter_id')) is True:
+            confirmview = confirm(ctx, self.client, 30.0)
+            confirmembed = discord.Embed(description=f"Are you sure you want to change the current entry number of **lottery #{lottery_id}** to `{count}`?\n\nThe last correctly entered entry will be `{count}`.\nThe next entry to be entered will be `{count+1}`, ", color=discord.Color.orange())
+            confirmview.response = await ctx.send(embed=confirmembed, view=confirmview)
+            await confirmview.wait()
+            if confirmview.returning_value is not True:
+                confirmembed.color = discord.Color.red()
+                return await confirmview.response.edit(embed=confirmembed)
+            else:
+                confirmembed.color = discord.Color.green()
+                confirmembed.description += f"\n\n<:DVB_True:887589686808309791> **Success!**"
+                await self.client.db.execute("DELETE FROM lottery_entries WHERE lottery_id = $1 AND lottery_number > $2", lottery_id, count)
             return await confirmview.response.edit(embed=confirmembed)
-        else:
-            confirmembed.color = discord.Color.green()
-            confirmembed.description += f"\n\n<:DVB_True:887589686808309791> **Success!**"
-            await self.client.db.execute("DELETE FROM lottery_entries WHERE lottery_id = $1 AND lottery_number > $2", lottery_id, count)
-        return await confirmview.response.edit(embed=confirmembed)
 
 
 
 
     @checks.has_permissions_or_role(manage_roles=True)
     @lottery.command(name='end')
-    async def lottery_end(self, ctx: DVVTcontext):
-        pass
+    async def lottery_end(self, ctx: DVVTcontext, lottery_id: int):
+        """
+        Ends a lottery.
+        """
+        if (lotto_object := await self.client.db.fetchrow("SELECT * FROM lotteries WHERE lottery_id = $1 AND active = $2", lottery_id, True)) is None:
+            return await ctx.send(f"<:DVB_False:887589731515392000> An active lottery with the ID {lottery_id} does not exist.")
+
+        if await self.check_non_hoster_consent(ctx, lotto_object.get('starter_id')) is True:
+            max_lotto_no = await self.client.db.fetchval("SELECT MAX(lottery_number) FROM lottery_entries WHERE lottery_id = $1", lottery_id)
+            confirmview = confirm(ctx, self.client, 30.0)
+            confirmembed = discord.Embed(title=f"Are you sure you want to end lottery #{lottery_id}?\n\nNumber of entries: {max_lotto_no}\n", color=discord.Color.orange())
+            confirmview.response = await ctx.send(embed=confirmembed, view=confirmview)
+            await confirmview.wait()
+            if confirmview.returning_value is not True:
+                confirmembed.color = discord.Color.red()
+                return await confirmview.response.edit(embed=confirmembed)
+            else:
+                confirmembed.color = discord.Color.green()
+                await confirmview.response.edit(embed=confirmembed)
+                if max_lotto_no < 1:
+                    the_chosen_one = 0
+                else:
+                    the_chosen_one = random.randint(1, max_lotto_no)
+                await self.client.db.execute("UPDATE lotteries SET active = $1 WHERE lottery_id = $2", False, lottery_id)
+                lottery_win_embed = discord.Embed(title=f"{max_lotto_no} users participated!", color=self.client.embed_color)
+                lottery_win_embed.set_footer(text=f"Lottery #{lottery_id} • {ctx.guild.name}", icon_url=ctx.guild.icon.url)
+                lottery_win_embed.set_thumbnail(url=ctx.guild.icon.url)
+                lottery_win_embed.add_field(name="Winning Ticket Number:", value=f"```\n{the_chosen_one}\n```")
+                if (channel := ctx.guild.get_channel(get_lotto_channel(ctx.guild.id, lotto_object.get('lottery_type')))) is not None:
+                    await channel.send(embed=lottery_win_embed)
+                with contextlib.suppress(discord.Forbidden):
+                    if the_chosen_one == 0:
+                        await ctx.author.send("Looks like no one joined your lottery.", view=LottoTemplateView(self.client, lotto_object.get('lottery_type'), 0, None))
+                    else:
+                        us = self.client.get_user(await self.client.db.fetchval("SELECT lottery_user FROM lottery_entries WHERE lottery_number = $1 AND lottery_id = $2", the_chosen_one, lottery_id))
+                        if lotto_object.get('lottery_type') == 'owo':
+                            msg = f"A winner for the lottery #{lottery_id} in {ctx.guild.name} has been chosen in <#859761455242149919>."
+                        elif lotto_object.get('lottery_type') == 'dank':
+                            msg = f"A winner for the lottery #{lottery_id} in {ctx.guild.name} has been chosen in <#731398659664511057>, Don't forget to echo the following text in <#680001973352923156>:"
+                        elif lotto_object.get('lottery_type') == 'karuta':
+                            msg = f"A winner for the lottery #{lottery_id} in {ctx.guild.name} has been chosen in <#887006001566462062>."
+                        else:
+                            msg = f"Generate a template to be sent to show the winner of this lottery."
+                        await ctx.author.send(msg, view=LottoTemplateView(self.client, lotto_object.get('lottery_type'), max_lotto_no, us))
